Add running state tracking to Clock class
The Clock class now includes a `running` field to track whether the clock is currently running. A new public property `Running` has been added to expose this state. The `StartAsync` method has been updated to set the `running` field to `true` when the clock starts and to `false` when the loop exits.